// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      titles: {
        Row: {
          alternative_titles: string[] | null
          artist: string | null
          author: string
          cover: string
          created_at: string
          genres: string[]
          id: string
          rating: number | null
          slug: string | null
          status: string
          synopsis: string
          title: string
          type: string
          updated_at: string
          views: number | null
          year: number
        }
        Insert: {
          alternative_titles?: string[] | null
          artist?: string | null
          author: string
          cover: string
          created_at?: string
          genres?: string[]
          id?: string
          rating?: number | null
          slug?: string | null
          status: string
          synopsis: string
          title: string
          type: string
          updated_at?: string
          views?: number | null
          year: number
        }
        Update: {
          alternative_titles?: string[] | null
          artist?: string | null
          author?: string
          cover?: string
          created_at?: string
          genres?: string[]
          id?: string
          rating?: number | null
          slug?: string | null
          status?: string
          synopsis?: string
          title?: string
          type?: string
          updated_at?: string
          views?: number | null
          year?: number
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never